requirements.txt
requests>=2.20.0,<3.0
PyJWT==2.6.0
cryptography==40.0.2